separate to own files instead of 3k lines in a single file
