+++ src/musicat/player.cpp
@@ -1,17 +1,6 @@
-#include <cstdio>
-#include <deque>
-#include <libpq-fe.h>
-#include <ogg/ogg.h>
-// #include <opus/opusfile.h>
-#include "musicat/cmds.h"
+#include "musicat/player.h"
 #include "musicat/db.h"
 #include "musicat/musicat.h"
-#include "musicat/player.h"
-#include <chrono>
-#include <dirent.h>
-#include <filesystem>
-#include <regex>
-#include <sys/stat.h>
 
 namespace musicat
 {
@@ -93,7 +82,9 @@ Player::add_track (MCTrack track, bool top, dpp::snowflake guild_id,
         if (top)
             {
                 this->queue.push_front (track);
-                if (this->shifted_track < /* Check queue size after push_front */ (this->queue.size () - 1))
+                if (this->shifted_track
+                    < /* Check queue size after push_front */ (
+                        this->queue.size () - 1))
                     {
                         std::lock_guard<std::mutex> lk (this->st_m);
                         this->shifted_track++;
@@ -345,1918 +336,6 @@ Player::is_stopped ()
 {
     std::lock_guard<std::mutex> lk (s_m);
     return this->stopped;
-}
-
-Manager::Manager (dpp::cluster *cluster, dpp::snowflake sha_id)
-{
-    this->cluster = cluster;
-    this->sha_id = sha_id;
-}
-
-Manager::~Manager () = default;
-
-std::shared_ptr<Player>
-Manager::create_player (dpp::snowflake guild_id)
-{
-    std::lock_guard<std::mutex> lk (this->ps_m);
-    auto l = players.find (guild_id);
-    if (l != players.end ())
-        return l->second;
-    std::shared_ptr<Player> v = std::make_shared<Player> (cluster, guild_id);
-    v->manager = this;
-    players.insert (std::pair (guild_id, v));
-    return v;
-}
-
-std::shared_ptr<Player>
-Manager::get_player (dpp::snowflake guild_id)
-{
-    std::lock_guard<std::mutex> lk (this->ps_m);
-    auto l = players.find (guild_id);
-    if (l != players.end ())
-        return l->second;
-    return NULL;
-}
-
-void
-Manager::reconnect (dpp::discord_client *from, dpp::snowflake guild_id)
-{
-    bool from_dc = false;
-    {
-        std::unique_lock<std::mutex> lk (this->dc_m);
-        auto a = this->disconnecting.find (guild_id);
-        if (a != this->disconnecting.end ())
-            {
-                from_dc = true;
-                this->dl_cv.wait (lk, [this, &guild_id] () {
-                    auto t = this->disconnecting.find (guild_id);
-                    return t == this->disconnecting.end ();
-                });
-            }
-    }
-    {
-        std::unique_lock<std::mutex> lk (this->c_m);
-        auto a = this->connecting.find (guild_id);
-        if (a != this->connecting.end ())
-            {
-                {
-                    using namespace std::chrono_literals;
-                    if (from_dc)
-                        std::this_thread::sleep_for (500ms);
-                }
-                from->connect_voice (guild_id, a->second, false, true);
-                this->dl_cv.wait (lk, [this, &guild_id] () {
-                    auto t = this->connecting.find (guild_id);
-                    return t == this->connecting.end ();
-                });
-            }
-    }
-}
-
-bool
-Manager::delete_player (dpp::snowflake guild_id)
-{
-    std::lock_guard<std::mutex> lk (this->ps_m);
-    auto l = players.find (guild_id);
-    if (l == players.end ())
-        return false;
-    players.erase (l);
-    return true;
-}
-
-std::deque<MCTrack>
-Manager::get_queue (dpp::snowflake guild_id)
-{
-    auto p = get_player (guild_id);
-    if (!p)
-        return {};
-    p->reset_shifted ();
-    return p->queue;
-}
-
-bool
-Manager::pause (dpp::discord_client *from, dpp::snowflake guild_id,
-                dpp::snowflake user_id)
-{
-    auto p = get_player (guild_id);
-    if (!p)
-        return false;
-    bool a = p->pause (from, user_id);
-    if (a)
-        {
-            std::lock_guard<std::mutex> lk (mp_m);
-            if (vector_find (&this->manually_paused, guild_id)
-                == this->manually_paused.end ())
-                this->manually_paused.push_back (guild_id);
-            this->update_info_embed (guild_id);
-        }
-    return a;
-}
-
-bool
-Manager::unpause (dpp::discord_voice_client *voiceclient,
-                  dpp::snowflake guild_id)
-{
-    std::lock_guard<std::mutex> lk (this->mp_m);
-    auto l = vector_find (&this->manually_paused, guild_id);
-    bool ret;
-    if (l != this->manually_paused.end ())
-        {
-            this->manually_paused.erase (l);
-            ret = true;
-        }
-    else
-        ret = false;
-    voiceclient->pause_audio (false);
-    this->update_info_embed (guild_id);
-    return ret;
-}
-
-bool
-Manager::is_disconnecting (dpp::snowflake guild_id)
-{
-    std::lock_guard<std::mutex> lk1 (this->dc_m);
-    return this->disconnecting.find (guild_id) != this->disconnecting.end ();
-}
-
-bool
-Manager::is_connecting (dpp::snowflake guild_id)
-{
-    std::lock_guard<std::mutex> lk2 (this->c_m);
-    return this->connecting.find (guild_id) != this->connecting.end ();
-}
-
-bool
-Manager::is_waiting_vc_ready (dpp::snowflake guild_id)
-{
-    std::lock_guard<std::mutex> lk3 (this->wd_m);
-    return this->waiting_vc_ready.find (guild_id)
-           != this->waiting_vc_ready.end ();
-}
-
-bool
-Manager::voice_ready (dpp::snowflake guild_id, dpp::discord_client *from,
-                      dpp::snowflake user_id)
-{
-    bool re = is_connecting (guild_id);
-    if (is_disconnecting (guild_id) || re || is_waiting_vc_ready (guild_id))
-        {
-            if (re && from)
-                {
-                    std::thread t (
-                        [this, user_id, guild_id] (dpp::discord_client *from) {
-                            bool user_vc = true;
-                            std::pair<
-                                dpp::channel *,
-                                std::map<dpp::snowflake, dpp::voicestate> >
-                                uservc;
-
-                            try
-                                {
-                                    uservc = get_voice_from_gid (guild_id,
-                                                                 user_id);
-                                }
-                            catch (...)
-                                {
-                                    user_vc = false;
-                                }
-
-                            try
-                                {
-                                    auto f = from->connecting_voice_channels
-                                                 .find (guild_id);
-                                    auto c = get_voice_from_gid (
-                                        guild_id, from->creator->me.id);
-                                    if (f
-                                            == from->connecting_voice_channels
-                                                   .end ()
-                                        || !f->second)
-                                        {
-                                            std::lock_guard<std::mutex> lk (
-                                                this->dc_m);
-                                            this->disconnecting[guild_id] = 1;
-                                            from->disconnect_voice (guild_id);
-                                        }
-                                    else if (user_vc
-                                             && uservc.first->id
-                                                    != c.first->id)
-                                        {
-                                            if (get_debug_state ())
-                                                printf (
-                                                    "Disconnecting as it "
-                                                    "seems I just got moved "
-                                                    "to different vc and "
-                                                    "connection not updated "
-                                                    "yet: %ld\n",
-                                                    guild_id);
-                                            std::lock_guard<std::mutex> lk (
-                                                this->dc_m);
-                                            std::lock_guard<std::mutex> lk2 (
-                                                this->c_m);
-                                            this->disconnecting[guild_id]
-                                                = f->second->channel_id;
-                                            this->connecting[guild_id]
-                                                = uservc.first->id;
-                                            from->disconnect_voice (guild_id);
-                                        }
-                                }
-                            catch (...)
-                                {
-                                    reset_voice_channel (from, guild_id);
-
-                                    if (user_id && user_vc)
-                                        try
-                                            {
-                                                std::lock_guard<std::mutex>
-                                                    lk (this->c_m);
-                                                auto p
-                                                    = this->connecting.find (
-                                                        guild_id);
-                                                std::map<dpp::snowflake,
-                                                         dpp::voicestate>
-                                                    vm = {};
-                                                if (p
-                                                    != this->connecting.end ())
-                                                    {
-                                                        auto gc = dpp::
-                                                            find_channel (
-                                                                p->second);
-                                                        if (gc)
-                                                            vm = gc->get_voice_members ();
-                                                        auto l = has_listener (
-                                                            &vm);
-                                                        if (!l
-                                                            && p->second
-                                                                   != uservc
-                                                                          .first
-                                                                          ->id)
-                                                            p->second
-                                                                = uservc.first
-                                                                      ->id;
-                                                    }
-                                            }
-                                        catch (...)
-                                            {
-                                            }
-                                }
-                            this->reconnect (from, guild_id);
-                        },
-                        from);
-                    t.detach ();
-                }
-            return false;
-        }
-    return true;
-}
-
-void
-Manager::stop_stream (dpp::snowflake guild_id)
-{
-    {
-        std::lock_guard<std::mutex> lk (this->sq_m);
-        if (vector_find (&this->stop_queue, guild_id)
-            == this->stop_queue.end ())
-            this->stop_queue.push_back (guild_id);
-    }
-}
-
-int
-Manager::skip (dpp::voiceconn *v, dpp::snowflake guild_id,
-               dpp::snowflake user_id, int64_t amount, bool remove)
-{
-    if (!v)
-        return -1;
-    auto p = get_player (guild_id);
-    if (!p)
-        return -1;
-    try
-        {
-            auto u = get_voice_from_gid (guild_id, user_id);
-            if (u.first->id != v->channel_id)
-                throw exception ("You're not in my voice channel", 0);
-
-            unsigned siz = 0;
-            for (auto &i : u.second)
-                {
-                    auto &a = i.second;
-                    if (a.is_deaf () || a.is_self_deaf ())
-                        continue;
-                    auto user = dpp::find_user (a.user_id);
-                    if (user->is_bot ())
-                        continue;
-                    siz++;
-                }
-            // if (siz > 1U)
-            // {
-            //     std::lock_guard<std::mutex> lk(p->q_m);
-            //     auto& track = p->queue.at(0);
-            //     if (track.user_id != user_id && track.user_id !=
-            //     this->sha_id)
-            //     {
-            //         amount = 1;
-            //         bool exist = false;
-            //         for (const auto& i : track.skip_vote)
-            //         {
-            //             if (i == user_id)
-            //             {
-            //                 exist = true;
-            //                 break;
-            //             }
-            //         }
-            //         if (!exist)
-            //         {
-            //             track.skip_vote.push_back(user_id);
-            //         }
-
-            //         unsigned ts = siz / 2U + 1U;
-            //         size_t ret = track.skip_vote.size();
-            //         if (ret < ts) return (int)ret;
-            //         else track.skip_vote.clear();
-            //     }
-            //     else if (amount > 1)
-            //     {
-            //         int64_t count = 0;
-            //         for (const auto& t : p->queue)
-            //         {
-            //             if (t.user_id == user_id || t.user_id ==
-            //             this->sha_id) count++; else break;
-
-            //             if (amount == count) break;
-            //         }
-            //         if (amount > count) amount = count;
-            //     }
-            // }
-        }
-    catch (const char *e)
-        {
-            throw exception ("You're not in a voice channel", 1);
-        }
-    auto siz = p->queue.size ();
-    if (amount < (siz || 1))
-        amount = siz || 1;
-    if (amount > 1000)
-        amount = 1000;
-    const bool l_s = p->loop_mode == loop_mode_t::l_song_queue;
-    const bool l_q = p->loop_mode == loop_mode_t::l_queue;
-    {
-        std::lock_guard<std::mutex> lk (p->q_m);
-        for (int64_t i = (p->loop_mode == loop_mode_t::l_song || l_s) ? 0 : 1;
-             i < amount; i++)
-            {
-                if (p->queue.begin () == p->queue.end ())
-                    break;
-                auto l = p->queue.front ();
-                p->queue.pop_front ();
-                if (!remove && (l_s || l_q))
-                    p->queue.push_back (l);
-            }
-    }
-    if (v && v->voiceclient && v->voiceclient->get_secs_remaining () > 0.1)
-        this->stop_stream (guild_id);
-
-    int a = p->skip (v);
-
-    if (v && v->voiceclient && remove)
-        v->voiceclient->insert_marker ("rm");
-
-    return a;
-}
-
-void
-Manager::download (string fname, string url, dpp::snowflake guild_id)
-{
-    std::thread tj (
-        [this] (string fname, string url, dpp::snowflake guild_id) {
-            {
-                std::lock_guard<std::mutex> lk (this->dl_m);
-                this->waiting_file_download[fname] = guild_id;
-            }
-
-            const std::string music_folder_path = get_music_folder_path ();
-
-            {
-                struct stat buf;
-                if (stat (music_folder_path.c_str (), &buf) != 0)
-                    std::filesystem::create_directory (music_folder_path);
-            }
-
-            string cmd
-                = string ("yt-dlp -f 251 --http-chunk-size 2M '") + url
-                  + string ("' -x --audio-format opus --audio-quality 0 -o '")
-                  + music_folder_path
-                  + std::regex_replace (fname, std::regex ("(')"), "'\\''",
-                                        std::regex_constants::match_any)
-                  + string ("'");
-
-            const bool debug = get_debug_state ();
-
-            if (debug)
-                {
-                    printf ("DOWNLOAD: \"%s\" \"%s\"\n", fname.c_str (),
-                            url.c_str ());
-                    printf ("CMD: %s\n", cmd.c_str ());
-                }
-            else
-                cmd += " 1>/dev/null";
-
-            system (cmd.c_str ());
-
-            {
-                std::lock_guard<std::mutex> lk (this->dl_m);
-                this->waiting_file_download.erase (fname);
-            }
-
-            this->dl_cv.notify_all ();
-        },
-        fname, url, guild_id);
-    tj.detach ();
-}
-
-void
-Manager::wait_for_download (string file_name)
-{
-    std::unique_lock<std::mutex> lk (this->dl_m);
-    this->dl_cv.wait (lk, [this, file_name] () {
-        return this->waiting_file_download.find (file_name)
-               == this->waiting_file_download.end ();
-    });
-}
-
-void
-Manager::stream (dpp::discord_voice_client *v, string fname)
-{
-    dpp::snowflake server_id;
-    std::chrono::_V2::system_clock::time_point start_time;
-
-    const string music_folder_path = get_music_folder_path ();
-
-    const bool debug = get_debug_state ();
-    if (v && !v->terminating && v->is_ready ())
-        {
-            FILE *fd;
-            ogg_sync_state oy;
-            ogg_stream_state os;
-            try
-                {
-                    server_id = v->server_id;
-
-                    const string file_path = music_folder_path + fname;
-
-                    start_time = std::chrono::high_resolution_clock::now ();
-
-                    if (debug)
-                        printf ("Streaming \"%s\" to %ld\n", fname.c_str (),
-                                server_id);
-
-                    fd = fopen (file_path.c_str (), "rb");
-                    if (!fd)
-                        {
-                            std::filesystem::create_directory (
-                                music_folder_path);
-                            throw 2;
-                        }
-
-                    struct stat buf;
-                    if (fstat (fileno (fd), &buf) != 0)
-                        {
-                            fclose (fd);
-                            throw 2;
-                        }
-
-                    ogg_page og;
-                    ogg_packet op;
-                    // OpusHead header;
-
-                    size_t sz = buf.st_size;
-                    if (debug)
-                        printf ("BUFFER_SIZE: %ld\n", sz);
-
-                    ogg_sync_init (&oy);
-
-                    // int eos = 0;
-                    // int i;
-
-                    fread (ogg_sync_buffer (&oy, sz), 1, sz, fd);
-                    fclose (fd);
-                    fd = NULL;
-
-                    bool no_prob = true;
-
-                    ogg_sync_wrote (&oy, sz);
-
-                    if (ogg_sync_pageout (&oy, &og) != 1)
-                        {
-                            fprintf (stderr,
-                                     "Does not appear to be ogg stream.\n");
-                            no_prob = false;
-                        }
-
-                    ogg_stream_init (&os, ogg_page_serialno (&og));
-
-                    if (ogg_stream_pagein (&os, &og) < 0)
-                        {
-                            fprintf (
-                                stderr,
-                                "Error reading initial page of ogg stream.\n");
-                            no_prob = false;
-                        }
-
-                    if (ogg_stream_packetout (&os, &op) != 1)
-                        {
-                            fprintf (stderr, "Error reading header packet of "
-                                             "ogg stream.\n");
-                            no_prob = false;
-                        }
-
-                    /* We must ensure that the ogg stream actually contains
-                     * opus data */
-                    // if (!(op.bytes > 8 && !memcmp("OpusHead", op.packet,
-                    // 8)))
-                    // {
-                    //     fprintf(stderr, "Not an ogg opus stream.\n");
-                    //     exit(1);
-                    // }
-
-                    // /* Parse the header to get stream info */
-                    // int err = opus_head_parse(&header, op.packet, op.bytes);
-                    // if (err)
-                    // {
-                    //     fprintf(stderr, "Not a ogg opus stream\n");
-                    //     exit(1);
-                    // }
-                    // /* Now we ensure the encoding is correct for Discord */
-                    // if (header.channel_count != 2 &&
-                    // header.input_sample_rate != 48000)
-                    // {
-                    //     fprintf(stderr, "Wrong encoding for Discord, must be
-                    //     48000Hz sample rate with 2 channels.\n"); exit(1);
-                    // }
-
-                    /* Now loop though all the pages and send the packets to
-                     * the vc */
-                    bool br = false;
-                    if (no_prob)
-                        while (true)
-                            {
-                                if (br)
-                                    {
-                                        if (debug)
-                                            printf ("[MANAGER::STREAM] "
-                                                    "Stopping stream\n");
-                                        break;
-                                    }
-
-                                {
-                                    std::lock_guard<std::mutex> lk (
-                                        this->sq_m);
-                                    auto sq = vector_find (&this->stop_queue,
-                                                           server_id);
-                                    if (sq != this->stop_queue.end ())
-                                        break;
-                                }
-
-                                if (!v || v->terminating)
-                                    {
-                                        fprintf (stderr,
-                                                 "[ERROR MANAGER::STREAM] "
-                                                 "Can't continue streaming, "
-                                                 "connection broken\n");
-                                        break;
-                                    }
-
-                                if (ogg_sync_pageout (&oy, &og) != 1)
-                                    {
-                                        // fprintf(stderr, "[ERROR
-                                        // MANAGER::STREAM] Can't continue
-                                        // streaming, corrupt audio (need
-                                        // recapture or incomplete audio
-                                        // file)\n");
-                                        break;
-                                    }
-
-                                if (ogg_stream_pagein (&os, &og) < 0)
-                                    {
-                                        fprintf (stderr,
-                                                 "[ERROR MANAGER::STREAM] "
-                                                 "Can't continue streaming, "
-                                                 "error reading page of Ogg "
-                                                 "bitstream data\n");
-                                        break;
-                                    }
-
-                                int po_res;
-                                while (
-                                    (po_res = ogg_stream_packetout (&os, &op)))
-                                    {
-                                        if (po_res == -1)
-                                            {
-                                                fprintf (
-                                                    stderr,
-                                                    "[WARN MANAGER::STREAM] "
-                                                    "Audio gap\n");
-                                            }
-                                        if (po_res == 0)
-                                            {
-                                                fprintf (
-                                                    stderr,
-                                                    "[WARN MANAGER::STREAM] "
-                                                    "Async or delayed "
-                                                    "audio\n");
-                                                break;
-                                            }
-
-                                        // /* Read remaining headers */
-                                        // if (op.bytes > 8 &&
-                                        // !memcmp("OpusHead", op.packet, 8))
-                                        // {
-                                        //     int err =
-                                        //     opus_head_parse(&header,
-                                        //     op.packet, op.bytes); if (err)
-                                        //     {
-                                        //         fprintf(stderr, "Not a ogg
-                                        //         opus stream\n"); exit(1);
-                                        //     }
-                                        //     if (header.channel_count != 2 &&
-                                        //     header.input_sample_rate !=
-                                        //     48000)
-                                        //     {
-                                        //         fprintf(stderr, "Wrong
-                                        //         encoding for Discord, must
-                                        //         be 48000Hz sample rate with
-                                        //         2 channels.\n"); exit(1);
-                                        //     }
-                                        //     continue;
-                                        // }
-                                        /* Skip the opus tags */
-                                        /* if (op.bytes > 8 &&
-                                         * !memcmp("OpusTags", op.packet, 8))
-                                         */
-                                        /* continue; */
-
-                                        /* Send the audio */
-                                        // int samples =
-                                        // opus_packet_get_samples_per_frame(op.packet,
-                                        // 48000);
-
-                                        if (v && !v->terminating)
-                                            v->send_audio_opus (
-                                                op.packet, op.bytes,
-                                                20); // samples / 48);
-                                        /* if (ogg_stream_eos(&os)) */
-                                        /*     break; */
-
-                                        br = v->terminating;
-
-                                        while (v && !v->terminating
-                                               && v->get_secs_remaining ()
-                                                      > 0.5)
-                                            {
-                                                std::this_thread::sleep_for (
-                                                    std::chrono::milliseconds (
-                                                        250));
-                                                if (v->terminating)
-                                                    br = true;
-                                            }
-
-                                        {
-                                            std::lock_guard<std::mutex> lk (
-                                                this->sq_m);
-                                            auto sq = vector_find (
-                                                &this->stop_queue, server_id);
-                                            if (sq != this->stop_queue.end ())
-                                                {
-                                                    br = true;
-                                                    break;
-                                                }
-                                        }
-
-                                        if (br)
-                                            {
-                                                if (debug)
-                                                    printf (
-                                                        "[MANAGER::STREAM] "
-                                                        "Stopping stream\n");
-                                                break;
-                                            }
-                                    }
-                            }
-                }
-            catch (const std::system_error &e)
-                {
-                    fprintf (stderr, "[ERROR MANAGER::STREAM] %d: %s\n",
-                             e.code ().value (), e.what ());
-                }
-            /* Cleanup */
-            ogg_stream_clear (&os);
-            ogg_sync_clear (&oy);
-            auto end_time = std::chrono::high_resolution_clock::now ();
-            auto done = std::chrono::duration_cast<std::chrono::milliseconds> (
-                end_time - start_time);
-            if (debug)
-                printf ("Done streaming for %ld milliseconds\n",
-                        done.count ());
-        }
-    else
-        throw 1;
-}
-
-void
-Manager::play (dpp::discord_voice_client *v, string fname,
-               dpp::snowflake channel_id, bool notify_error)
-{
-    std::thread tj (
-        [this] (dpp::discord_voice_client *v, string fname,
-                dpp::snowflake channel_id, bool notify_error) {
-            const bool debug = get_debug_state ();
-
-            if (debug)
-                printf ("Attempt to stream\n");
-            auto server_id = v->server_id;
-            auto voice_channel_id = v->channel_id;
-
-            try
-                {
-                    this->stream (v, fname);
-                }
-            catch (int e)
-                {
-                    fprintf (stderr,
-                             "[ERROR MANAGER::PLAY] Stream thrown error with "
-                             "code: %d\n",
-                             e);
-
-                    if (notify_error)
-                        {
-                            string msg = "";
-
-                            // Maybe connect/reconnect here if there's
-                            // connection error
-                            if (e == 2)
-                                msg = "Can't start playback";
-                            else if (e == 1)
-                                msg = "No connection";
-
-                            dpp::message m;
-                            m.set_channel_id (channel_id).set_content (msg);
-
-                            this->cluster->message_create (m);
-                        }
-                }
-
-            if (server_id)
-                {
-                    std::lock_guard<std::mutex> lk (this->sq_m);
-
-                    auto sq = vector_find (&this->stop_queue, server_id);
-                    if (sq != this->stop_queue.end ())
-                        {
-                            if (debug)
-                                printf ("[MANAGER::STREAM] Stopped because "
-                                        "stop query, cleaning up query\n");
-                            this->stop_queue.erase (sq);
-                            this->stop_queue_cv.notify_all ();
-                        }
-                }
-
-            if (v && !v->terminating)
-                v->insert_marker ("e");
-            else
-                {
-                    try
-                        {
-                            get_voice_from_gid (server_id, this->sha_id);
-                            return;
-                        }
-                    catch (...)
-                        {
-                        }
-
-                    if (server_id && voice_channel_id)
-                        {
-                            std::lock_guard<std::mutex> lk (this->c_m);
-                            this->connecting[server_id] = voice_channel_id;
-                        }
-                    // if (v) v->~discord_voice_client();
-                }
-        },
-        v, fname, channel_id, notify_error);
-    tj.detach ();
-}
-
-bool
-Manager::send_info_embed (dpp::snowflake guild_id, bool update,
-                          bool force_playing_status)
-{
-    {
-        auto player = this->get_player (guild_id);
-        if (!player)
-            throw exception ("No player");
-
-        const bool debug = get_debug_state ();
-
-        if (update && !player->info_message)
-            {
-                if (debug)
-                    printf (
-                        "[MANAGER:SEND_INFO_EMBED] No message to update\n");
-                return false;
-            }
-
-        auto channel_id = player->channel_id;
-
-        auto g = dpp::find_guild (guild_id);
-        auto c = dpp::find_channel (channel_id);
-        bool embed_perms = has_permissions (
-            g, &this->cluster->me, c,
-            { dpp::p_view_channel, dpp::p_send_messages, dpp::p_embed_links });
-        bool view_history_perm = has_permissions (
-            g, &this->cluster->me, c, { dpp::p_read_message_history });
-
-        bool delete_original = false;
-        if (update)
-            {
-                if (!view_history_perm
-                    || (player->info_message
-                        && player->info_message->is_source_message_deleted ()))
-                    {
-                        update = false;
-                        delete_original = true;
-                    }
-            }
-        else
-            {
-                if (!embed_perms)
-                    throw exception ("No permission");
-            }
-
-        dpp::embed e;
-        try
-            {
-                e = get_playing_info_embed (guild_id, force_playing_status);
-            }
-        catch (const exception &e)
-            {
-                fprintf (stderr,
-                         "[ERROR MANAGER::SEND_INFO_EMBED] Failed to "
-                         "get_playing_info_embed: %s\n",
-                         e.what ());
-                return false;
-            }
-        catch (const dpp::exception &e)
-            {
-                fprintf (stderr,
-                         "[ERROR MANAGER::SEND_INFO_EMBED] Failed to "
-                         "get_playing_info_embed [dpp::exception]: %s\n",
-                         e.what ());
-                return false;
-            }
-        catch (const std::logic_error &e)
-            {
-                fprintf (stderr,
-                         "[ERROR MANAGER::SEND_INFO_EMBED] Failed to "
-                         "get_playing_info_embed [std::logic_error]: %s\n",
-                         e.what ());
-                return false;
-            }
-
-        auto m_cb = [this, player, guild_id, channel_id,
-                     debug] (dpp::confirmation_callback_t cb) {
-            if (cb.is_error ())
-                {
-                    fprintf (stderr,
-                             "[ERROR MANAGER::SEND_INFO_EMBED] message_create "
-                             "callback error:\nmes: %s\ncode: %d\nerr:\n",
-                             cb.get_error ().message.c_str (),
-                             cb.get_error ().code);
-                    for (const auto &i : cb.get_error ().errors)
-                        fprintf (stderr, "c: %s\nf: %s\no: %s\nr: %s\n",
-                                 i.code.c_str (), i.field.c_str (),
-                                 i.object.c_str (), i.reason.c_str ());
-                    return;
-                }
-
-            if (cb.value.index ())
-                {
-                    if (!player)
-                        {
-                            fprintf (stderr,
-                                     "[ERROR MANAGER::SEND_INFO_EMBED] PLAYER "
-                                     "GONE WTFF\n");
-                            return;
-                        }
-
-                    std::lock_guard<std::mutex> lk (this->imc_m);
-                    if (player->info_message)
-                        {
-                            auto id = player->info_message->id;
-                            this->info_messages_cache.erase (id);
-                        }
-
-                    player->info_message = std::make_shared<dpp::message> (
-                        std::get<dpp::message> (cb.value));
-                    if (player->info_message)
-                        {
-                            this->info_messages_cache[player->info_message->id]
-                                = player->info_message;
-                            if (debug)
-                                printf ("[MANAGER::SEND_INFO_EMBED] New "
-                                        "message info: %ld\n",
-                                        player->info_message->id);
-                        }
-                }
-            else if (debug)
-                printf (
-                    "[MANAGER::SEND_INFO_EMBED] No message_create cb size\n");
-        };
-
-        if (delete_original)
-            {
-                this->delete_info_embed (guild_id);
-            }
-
-        // TODO: Refactor this horrendous `update` flag system and check for
-        // existing info_message instead
-        if (!update)
-            {
-                dpp::message m;
-                m.add_embed (e).set_channel_id (channel_id);
-
-                this->cluster->message_create (m, m_cb);
-            }
-        else if (player->info_message)
-            {
-                auto mn = *player->info_message;
-                if (!mn.embeds.empty ())
-                    mn.embeds.pop_back ();
-                mn.embeds.push_back (e);
-
-                if (debug)
-                    printf ("[MANAGER::SEND_INFO_EMBED] Channel Info Embed Id "
-                            "Edit: %ld %ld\n",
-                            mn.channel_id, mn.id);
-                this->cluster->message_edit (mn, m_cb);
-            }
-        return true;
-    }
-}
-
-bool
-Manager::update_info_embed (dpp::snowflake guild_id, bool force_playing_status)
-{
-    if (get_debug_state ())
-        printf ("[MANAGER::UPDATE_INFO_EMBED] Update info called\n");
-    return this->send_info_embed (guild_id, true, force_playing_status);
-}
-
-bool
-Manager::delete_info_embed (dpp::snowflake guild_id,
-                            dpp::command_completion_event_t callback)
-{
-    auto player = this->get_player (guild_id);
-    if (!player)
-        return false;
-
-    auto retdel = [player, this] () {
-        std::lock_guard<std::mutex> lk (this->imc_m);
-        if (player->info_message)
-            {
-                auto id = player->info_message->id;
-                this->info_messages_cache.erase (id);
-                return true;
-            }
-        else
-            return false;
-    };
-
-    if (!player->info_message)
-        return false;
-    if (player->info_message->is_source_message_deleted ())
-        return retdel ();
-
-    auto mid = player->info_message->id;
-    auto cid = player->info_message->channel_id;
-
-    if (get_debug_state ())
-        printf (
-            "[MANAGER::UPDATE_INFO_EMBED] Channel Info Embed Id Delete: %ld\n",
-            cid);
-    this->cluster->message_delete (mid, cid, callback);
-    return retdel ();
-}
-
-bool
-Manager::handle_on_track_marker (const dpp::voice_track_marker_t &event,
-                                 std::shared_ptr<Manager> shared_manager)
-{
-    const bool debug = get_debug_state ();
-    if (debug)
-        printf ("Handling voice marker: \"%s\" in guild %ld\n",
-                event.track_meta.c_str (), event.voice_client->server_id);
-
-    {
-        std::lock_guard<std::mutex> lk (this->mp_m);
-        auto l = vector_find (&this->manually_paused,
-                              event.voice_client->server_id);
-        if (l != this->manually_paused.end ())
-            {
-                this->manually_paused.erase (l);
-            }
-    }
-
-    if (!event.voice_client)
-        {
-            printf ("NO CLIENT\n");
-            return false;
-        }
-    if (this->is_disconnecting (event.voice_client->server_id))
-        {
-            if (debug)
-                printf ("RETURN DISCONNECTING\n");
-            return false;
-        }
-    auto p = this->get_player (event.voice_client->server_id);
-    if (!p)
-        {
-            if (debug)
-                printf ("NO PLAYER\n");
-            return false;
-        }
-
-    bool just_loaded_queue = false;
-    if (p->saved_queue_loaded != true)
-        {
-            this->load_guild_current_queue (event.voice_client->server_id,
-                                            &sha_id);
-            just_loaded_queue = true;
-        }
-    if (p->saved_config_loaded != true)
-        this->load_guild_player_config (event.voice_client->server_id);
-
-    MCTrack s;
-    {
-        {
-            std::lock_guard<std::mutex> lk (p->q_m);
-            if (p->queue.size () == 0)
-                {
-                    if (debug)
-                        printf ("NO SIZE BEFORE: %d\n", p->loop_mode);
-                    return false;
-                }
-        }
-
-        // Handle shifted tracks (tracks shifted to the front of the queue)
-        if (debug)
-            printf ("Resetting shifted: %d\n", p->reset_shifted ());
-        std::lock_guard<std::mutex> lk (p->q_m);
-
-        // Do stuff according to loop mode when playback ends
-        if (event.track_meta == "e" && !p->is_stopped ())
-            {
-                if (p->loop_mode == loop_mode_t::l_none)
-                    p->queue.pop_front ();
-                else if (p->loop_mode == loop_mode_t::l_queue)
-                    {
-                        auto l = p->queue.front ();
-                        p->queue.pop_front ();
-                        p->queue.push_back (l);
-                    }
-            }
-        else if (event.track_meta == "rm")
-            {
-                p->queue.pop_front ();
-                return false;
-            }
-
-        if (p->queue.size () == 0)
-            {
-                if (debug)
-                    printf ("NO SIZE AFTER: %d\n", p->loop_mode);
-                if (!just_loaded_queue)
-                    database::delete_guild_current_queue (
-                        event.voice_client->server_id);
-                return false;
-            }
-
-        p->queue.front ().skip_vote.clear ();
-        s = p->queue.front ();
-        p->set_stopped (false);
-        if (!just_loaded_queue)
-            database::update_guild_current_queue (
-                event.voice_client->server_id, p->queue);
-    }
-
-    try
-        {
-            auto c = get_voice_from_gid (event.voice_client->server_id,
-                                         this->sha_id);
-            if (!has_listener (&c.second))
-                return false;
-        }
-    catch (...)
-        {
-        }
-
-    if (event.voice_client && event.voice_client->get_secs_remaining () < 0.1)
-        {
-            std::thread tj (
-                [this, shared_manager,
-                 debug] (dpp::discord_voice_client *v, MCTrack track,
-                         string meta, std::shared_ptr<Player> player) {
-                    bool timed_out = false;
-                    auto guild_id = v->server_id;
-                    dpp::snowflake channel_id = player->channel_id;
-                    // std::thread tmt([this](bool* _v) {
-                    //     int _w = 30;
-                    //     while (_v && *_v == false && _w > 0)
-                    //     {
-                    //         sleep(1);
-                    //         --_w;
-                    //     }
-                    //     if (_w) return;
-                    //     if (_v)
-                    //     {
-                    //         *_v = true;
-                    //         this->dl_cv.notify_all();
-                    //     }
-                    // }, &timed_out);
-                    // tmt.detach();
-
-                    {
-                        std::unique_lock<std::mutex> lk (this->wd_m);
-                        auto a = this->waiting_vc_ready.find (guild_id);
-                        if (a != this->waiting_vc_ready.end ())
-                            {
-                                if (debug)
-                                    printf ("Waiting for ready state\n");
-
-                                this->dl_cv.wait (lk, [this, &guild_id,
-                                                       &timed_out] () {
-                                    auto t = this->waiting_vc_ready.find (
-                                        guild_id);
-                                    // if (timed_out)
-                                    // {
-                                    //     this->waiting_vc_ready.erase(t);
-                                    //     return true;
-                                    // }
-                                    auto c
-                                        = t == this->waiting_vc_ready.end ();
-                                    return c;
-                                });
-                            }
-                    }
-
-                    {
-                        std::unique_lock<std::mutex> lk (this->dl_m);
-                        auto a = this->waiting_file_download.find (
-                            track.filename);
-                        if (a != this->waiting_file_download.end ())
-                            {
-                                if (debug)
-                                    printf ("Waiting for download\n");
-
-                                this->dl_cv.wait (lk, [this, track,
-                                                       &timed_out] () {
-                                    auto t = this->waiting_file_download.find (
-                                        track.filename);
-                                    // if (timed_out)
-                                    // {
-                                    //     this->waiting_file_download.erase(t);
-                                    //     return true;
-                                    // }
-                                    auto c = t
-                                             == this->waiting_file_download
-                                                    .end ();
-                                    return c;
-                                });
-                            }
-                    }
-
-                    string id = track.id ();
-                    if (player->auto_play)
-                        {
-                            if (debug)
-                                printf ("Getting new autoplay track: %s\n",
-                                        id.c_str ());
-                            command::play::add_track (
-                                true, v->server_id,
-                                string ("https://www.youtube.com/watch?v=")
-                                    + id + "&list=RD" + id,
-                                0, true, NULL, 0, this->sha_id, shared_manager,
-                                false, player->from);
-                        }
-
-                    {
-                        std::lock_guard<std::mutex> lk (player->h_m);
-                        if (player->max_history_size)
-                            {
-                                player->history.push_back (id);
-                                while (player->history.size ()
-                                       > player->max_history_size)
-                                    {
-                                        player->history.pop_front ();
-                                    }
-                            }
-                    }
-
-                    auto c = dpp::find_channel (channel_id);
-                    auto g = dpp::find_guild (guild_id);
-                    bool embed_perms = has_permissions (
-                        g, &this->cluster->me, c,
-                        { dpp::p_view_channel, dpp::p_send_messages,
-                          dpp::p_embed_links });
-                    {
-                        const string music_folder_path = get_music_folder_path ();
-                        std::ifstream test (music_folder_path + track.filename,
-                                            std::ios_base::in
-                                                | std::ios_base::binary);
-                        if (!test.is_open ())
-                            {
-                                if (v && !v->terminating)
-                                    v->insert_marker ("e");
-                                if (embed_perms)
-                                    {
-                                        dpp::message m;
-                                        m.set_channel_id (channel_id)
-                                            .set_content ("Can't play track: "
-                                                          + track.title ()
-                                                          + " (added by <@"
-                                                          + std::to_string (
-                                                              track.user_id)
-                                                          + ">)");
-                                        this->cluster->message_create (m);
-                                    }
-                                return;
-                            }
-                        else
-                            test.close ();
-                    }
-                    if (timed_out)
-                        throw exception ("Operation took too long, aborted...",
-                                         0);
-                    if (meta == "r")
-                        v->send_silence (60);
-
-                    // Send play info embed
-                    try
-                        {
-                            this->play (v, track.filename, channel_id,
-                                        embed_perms);
-                            if (embed_perms)
-                                {
-                                    // Update if last message is the info embed
-                                    // message
-                                    if (c && player->info_message
-                                        && c->last_message_id
-                                        && c->last_message_id
-                                               == player->info_message->id)
-                                        {
-                                            if (player->loop_mode
-                                                    != loop_mode_t::l_song
-                                                && player->loop_mode
-                                                       != loop_mode_t::
-                                                           l_song_queue)
-                                                this->update_info_embed (
-                                                    guild_id, true);
-                                        }
-                                    else
-                                        {
-                                            this->delete_info_embed (guild_id);
-                                            this->send_info_embed (
-                                                guild_id, false, true);
-                                        }
-                                }
-                            else
-                                fprintf (stderr,
-                                         "[EMBED_UPDATE] No channel or "
-                                         "permission to send info embed\n");
-                        }
-                    catch (const exception &e)
-                        {
-                            fprintf (stderr, "[ERROR EMBED_UPDATE] %s\n",
-                                     e.what ());
-                            auto cd = e.code ();
-                            if (cd == 1 || cd == 2)
-                                if (embed_perms)
-                                    {
-                                        dpp::message m;
-                                        m.set_channel_id (channel_id)
-                                            .set_content (e.what ());
-                                        this->cluster->message_create (m);
-                                    }
-                        }
-                },
-                event.voice_client, s, event.track_meta, p);
-            tj.detach ();
-            return true;
-        }
-    else if (debug)
-        printf ("RETURN NO TRACK SIZE\n");
-    return false;
-}
-
-dpp::embed
-Manager::get_playing_info_embed (dpp::snowflake guild_id,
-                                 bool force_playing_status)
-{
-    auto player = this->get_player (guild_id);
-    if (!player)
-        throw exception ("No player");
-
-    // Reset shifted tracks
-    player->reset_shifted ();
-    MCTrack track;
-    MCTrack prev_track;
-    MCTrack next_track;
-    MCTrack skip_track;
-
-    {
-        std::lock_guard<std::mutex> lk (player->q_m);
-        auto siz = player->queue.size ();
-        if (!siz)
-            throw exception ("No track");
-        track = player->queue.front ();
-
-        prev_track = player->queue.at (siz - 1UL);
-
-        auto lm = player->loop_mode;
-        if (lm == loop_mode_t::l_queue)
-            next_track = (siz == 1UL) ? track : player->queue.at (1);
-        else if (lm == loop_mode_t::l_none)
-            {
-                if (siz > 1UL)
-                    next_track = player->queue.at (1);
-            }
-        else
-            {
-                // if (loop mode == one | one/queue)
-                next_track = track;
-                if (siz > 1UL)
-                    skip_track = player->queue.at (1);
-            }
-    }
-
-    dpp::guild_member o
-        = dpp::find_guild_member (guild_id, this->cluster->me.id);
-    dpp::guild_member u = dpp::find_guild_member (guild_id, track.user_id);
-    dpp::user *uc = dpp::find_user (u.user_id);
-    uint32_t color = 0;
-    for (auto i : o.roles)
-        {
-            auto r = dpp::find_role (i);
-            if (!r || !r->colour)
-                continue;
-            color = r->colour;
-            break;
-        }
-
-    string eaname = u.nickname.length () ? u.nickname : uc->username;
-    dpp::embed_author ea;
-    ea.name = eaname;
-    auto ua = u.get_avatar_url (4096);
-    auto uca = uc->get_avatar_url (4096);
-    if (ua.length ())
-        ea.icon_url = ua;
-    else if (uca.length ())
-        ea.icon_url = uca;
-
-    static const char *l_mode[]
-        = { "Repeat one", "Repeat queue", "Repeat one/queue" };
-    static const char *p_mode[] = { "Paused", "Playing" };
-
-    string et = track.bestThumbnail ().url;
-    dpp::embed e;
-    e.set_description (track.snippetText ())
-        .set_title (track.title ())
-        .set_url (track.url ())
-        .set_author (ea);
-
-    if (!prev_track.raw.is_null ())
-        e.add_field (
-            "PREVIOUS",
-            "[" + prev_track.title () + "](" + prev_track.url () + ")", true);
-    if (!next_track.raw.is_null ())
-        e.add_field (
-            "NEXT", "[" + next_track.title () + "](" + next_track.url () + ")",
-            true);
-    if (!skip_track.raw.is_null ())
-        e.add_field ("SKIP", "[" + skip_track.title () + "]("
-                                 + skip_track.url () + ")");
-
-    string ft = "";
-
-    bool tinfo = !track.info.raw.is_null ();
-    if (tinfo)
-        {
-            ft += format_duration (track.info.duration ());
-        }
-
-    bool has_p = false;
-
-    if (player->from)
-        {
-            auto con = player->from->get_voice (guild_id);
-            if (con && con->voiceclient)
-                if (con->voiceclient->get_secs_remaining () > 0.1)
-                    {
-                        has_p = true;
-                        if (ft.length ())
-                            ft += " | ";
-                        if (con->voiceclient->is_paused ())
-                            ft += p_mode[0];
-                        else
-                            ft += p_mode[1];
-                    }
-        }
-    if (force_playing_status && !has_p)
-        {
-            if (ft.length ())
-                ft += " | ";
-            ft += p_mode[1];
-        }
-
-    if (player->loop_mode)
-        {
-            if (ft.length ())
-                ft += " | ";
-            ft += l_mode[player->loop_mode - 1];
-        }
-    if (player->auto_play)
-        {
-            if (ft.length ())
-                ft += " | ";
-            ft += "Autoplay";
-            if (player->max_history_size)
-                ft += string (" (") + std::to_string (player->max_history_size)
-                      + ")";
-        }
-    if (tinfo)
-        {
-            if (ft.length ())
-                ft += " | ";
-            ft += string ("[") + std::to_string (track.info.average_bitrate ())
-                  + "]";
-        }
-    if (ft.length ())
-        e.set_footer (ft, "");
-    if (color)
-        e.set_color (color);
-    if (et.length ())
-        e.set_image (et);
-    return e;
-}
-
-std::vector<std::string>
-Manager::get_available_tracks (const size_t amount) const
-{
-    std::vector<std::string> ret = {};
-    size_t c = 0;
-    auto dir = opendir (get_music_folder_path().c_str());
-
-    if (dir != NULL)
-        {
-            auto file = readdir (dir);
-            while (file != NULL)
-                {
-                    if (file->d_type == DT_REG)
-                        {
-                            string s = string (file->d_name);
-                            ret.push_back (s.substr (0, s.length () - 5));
-                        }
-
-                    if (amount && ++c == amount)
-                        break;
-                    file = readdir (dir);
-                }
-            closedir (dir);
-        }
-    return ret;
-}
-
-void
-Manager::handle_on_voice_ready (const dpp::voice_ready_t &event)
-{
-    const bool debug = get_debug_state ();
-    {
-        std::lock_guard<std::mutex> lk (this->wd_m);
-        if (debug)
-            printf ("[EVENT] on_voice_ready\n");
-        if (this->waiting_vc_ready.find (event.voice_client->server_id)
-            != this->waiting_vc_ready.end ())
-            {
-                this->waiting_vc_ready.erase (event.voice_client->server_id);
-                this->dl_cv.notify_all ();
-            }
-    }
-
-    {
-        std::lock_guard<std::mutex> lk (this->c_m);
-        if (this->connecting.find (event.voice_client->server_id)
-            != this->connecting.end ())
-            {
-                this->connecting.erase (event.voice_client->server_id);
-                this->dl_cv.notify_all ();
-            }
-    }
-
-    auto i = event.voice_client->get_tracks_remaining ();
-    auto l = event.voice_client->get_secs_remaining ();
-    if (debug)
-        printf ("TO INSERT %d::%f\n", i, l);
-    if (l < 0.1)
-        {
-            event.voice_client->insert_marker ("r");
-            if (debug)
-                printf ("INSERTED \"r\" MARKER\n");
-        }
-}
-
-void
-Manager::handle_on_voice_state_update (const dpp::voice_state_update_t &event)
-{
-    const bool debug = get_debug_state ();
-
-    // Non client's user code
-    if (event.state.user_id != sha_id)
-        {
-            dpp::voiceconn *v = event.from->get_voice (event.state.guild_id);
-
-            if (!v || !v->channel_id || !v->voiceclient)
-                return;
-
-            // Pause audio when no user listening in vc
-            if (v->channel_id != event.state.channel_id
-                && v->voiceclient->get_secs_remaining () > 0.1
-                && !v->voiceclient->is_paused ())
-                {
-                    std::lock_guard<std::mutex> lk (this->mp_m);
-                    if (event.from
-                        && vector_find (&this->manually_paused,
-                                        event.state.guild_id)
-                               == this->manually_paused.end ())
-                        {
-                            try
-                                {
-                                    auto voice = get_voice_from_gid (
-                                        event.state.guild_id, sha_id);
-                                    // Whether there's human listening in the
-                                    // vc
-                                    bool p = false;
-                                    for (auto l : voice.second)
-                                        {
-                                            // This only check user in cache,
-                                            // if user not in cache then skip
-                                            auto a = dpp::find_user (l.first);
-                                            if (a)
-                                                {
-                                                    if (!a->is_bot ())
-                                                        {
-                                                            p = true;
-                                                            break;
-                                                        }
-                                                }
-                                        }
-                                    if (!p)
-                                        {
-                                            v->voiceclient->pause_audio (true);
-                                            this->update_info_embed (
-                                                event.state.guild_id);
-                                            if (debug)
-                                                printf ("Paused %ld as no "
-                                                        "user in vc\n",
-                                                        event.state.guild_id);
-                                        }
-                                }
-                            catch (const char *e)
-                                {
-                                    fprintf (stderr,
-                                             "[ERROR VOICE_STATE_UPDATE] %s\n",
-                                             e);
-                                }
-                        }
-                }
-            else
-                {
-                    if (v->channel_id == event.state.channel_id
-                        && v->voiceclient->get_secs_remaining () > 0.1
-                        && v->voiceclient->is_paused ())
-                        {
-                            std::lock_guard<std::mutex> lk (this->mp_m);
-                            // Whether the track paused automatically
-                            if (vector_find (&this->manually_paused,
-                                             event.state.guild_id)
-                                == this->manually_paused.end ())
-                                {
-                                    std::thread tj (
-                                        [this, event] (
-                                            dpp::discord_voice_client *vc) {
-                                            std::this_thread::sleep_for (
-                                                std::chrono::milliseconds (
-                                                    2500));
-                                            try
-                                                {
-                                                    auto a
-                                                        = get_voice_from_gid (
-                                                            event.state
-                                                                .guild_id,
-                                                            event.state
-                                                                .user_id);
-                                                    if (vc && !vc->terminating
-                                                        && a.first
-                                                        && a.first->id
-                                                               == event.state
-                                                                      .channel_id)
-                                                        {
-                                                            vc->pause_audio (
-                                                                false);
-                                                            this->update_info_embed (
-                                                                event.state
-                                                                    .guild_id);
-                                                        }
-                                                }
-                                            catch (...)
-                                                {
-                                                }
-                                        },
-                                        v->voiceclient);
-                                    tj.detach ();
-                                }
-                        }
-                    else if (event.state.channel_id
-                             && v->channel_id != event.state.channel_id)
-                        {
-                            try
-                                {
-                                    auto m = get_voice_from_gid (
-                                        event.state.guild_id, this->sha_id);
-                                    if (has_listener (&m.second))
-                                        {
-                                            std::lock_guard<std::mutex> lk (
-                                                this->dc_m);
-                                            std::lock_guard<std::mutex> lk2 (
-                                                this->c_m);
-                                            std::lock_guard<std::mutex> lk3 (
-                                                this->wd_m);
-                                            this->disconnecting[event.state
-                                                                    .guild_id]
-                                                = v->channel_id;
-                                            this->connecting[event.state
-                                                                 .guild_id]
-                                                = event.state.channel_id;
-                                            this->waiting_vc_ready
-                                                [event.state.guild_id]
-                                                = "0";
-                                            event.from->disconnect_voice (
-                                                event.state.guild_id);
-                                        }
-                                    std::thread tj ([this, event] () {
-                                        this->reconnect (event.from,
-                                                         event.state.guild_id);
-                                    });
-                                    tj.detach ();
-                                }
-                            catch (...)
-                                {
-                                }
-                        }
-                }
-
-            // End non client's user code
-            return;
-        }
-
-    // Client user code -----------------------------------
-
-    if (!event.state.channel_id)
-        {
-            std::lock_guard<std::mutex> lk (this->dc_m);
-            if (debug)
-                printf ("[EVENT] on_voice_state_leave\n");
-            if (this->disconnecting.find (event.state.guild_id)
-                != this->disconnecting.end ())
-                {
-                    this->disconnecting.erase (event.state.guild_id);
-                    this->dl_cv.notify_all ();
-                }
-        }
-    else
-        {
-            {
-                std::lock_guard<std::mutex> lk (this->c_m);
-                if (debug)
-                    printf ("[EVENT] on_voice_state_join\n");
-                if (this->connecting.find (event.state.guild_id)
-                    != this->connecting.end ())
-                    {
-                        this->connecting.erase (event.state.guild_id);
-                        this->dl_cv.notify_all ();
-                    }
-            }
-
-            try
-                {
-                    auto a = get_voice_from_gid (event.state.guild_id,
-                                                 event.state.user_id);
-                    auto v = event.from->get_voice (event.state.guild_id);
-
-                    if (v && v->voiceclient)
-                        {
-                            std::lock_guard<std::mutex> lk (this->wd_m);
-                            auto i = this->waiting_vc_ready.find (
-                                v->voiceclient->server_id);
-                            if (v->voiceclient->is_ready ()
-                                && i != this->waiting_vc_ready.end ())
-                                {
-                                    this->waiting_vc_ready.erase (i);
-                                    this->dl_cv.notify_all ();
-                                }
-                        }
-
-                    if (v && v->channel_id && v->channel_id != a.first->id)
-                        {
-                            this->stop_stream (event.state.guild_id);
-
-                            if (v->voiceclient && !v->voiceclient->terminating)
-                                {
-                                    v->voiceclient->pause_audio (false);
-                                    v->voiceclient->skip_to_next_marker ();
-                                }
-
-                            {
-                                std::lock_guard<std::mutex> lk (this->dc_m);
-                                this->disconnecting[event.state.guild_id]
-                                    = v->channel_id;
-                                event.from->disconnect_voice (
-                                    event.state.guild_id);
-                            }
-
-                            if (has_listener (&a.second))
-                                {
-                                    std::lock_guard<std::mutex> lk (this->c_m);
-                                    std::lock_guard<std::mutex> lk2 (
-                                        this->wd_m);
-                                    this->connecting.insert_or_assign (
-                                        event.state.guild_id, a.first->id);
-                                    this->waiting_vc_ready[event.state
-                                                               .guild_id]
-                                        = "0";
-                                }
-
-                            std::thread tj ([this, event] () {
-                                std::this_thread::sleep_for (
-                                    std::chrono::seconds (1));
-                                this->voice_ready (event.state.guild_id,
-                                                   event.from);
-                            });
-                            tj.detach ();
-                        }
-                }
-            catch (...)
-                {
-                }
-        }
-    // if (muted) player_manager->pause(event.guild_id);
-    // else player_manager->resume(guild_id);
-}
-
-bool
-Manager::set_info_message_as_deleted (dpp::snowflake id)
-{
-    std::lock_guard<std::mutex> lk (this->imc_m);
-    auto m = this->info_messages_cache.find (id);
-    if (m != this->info_messages_cache.end ())
-        {
-            if (!m->second->is_source_message_deleted ())
-                {
-                    m->second->set_flags (m->second->flags
-                                          | dpp::m_source_message_deleted);
-                    return true;
-                }
-        }
-    return false;
-}
-
-void
-Manager::handle_on_message_delete (const dpp::message_delete_t &event)
-{
-    this->set_info_message_as_deleted (event.deleted->id);
-}
-
-void
-Manager::handle_on_message_delete_bulk (
-    const dpp::message_delete_bulk_t &event)
-{
-    for (auto i : event.deleted)
-        this->set_info_message_as_deleted (i);
-}
-
-size_t
-Manager::remove_track (dpp::snowflake guild_id, size_t pos,
-                       const size_t amount, const size_t to)
-{
-    auto p = this->get_player (guild_id);
-    if (!p)
-        return 0;
-    return p->remove_track (pos, amount, to);
-}
-
-bool
-Manager::shuffle_queue (dpp::snowflake guild_id)
-{
-    auto p = this->get_player (guild_id);
-    if (!p)
-        return false;
-    return p->shuffle ();
-}
-
-void
-Manager::set_ignore_marker (const dpp::snowflake &guild_id)
-{
-    std::lock_guard<std::mutex> lk (this->im_m);
-    auto l = vector_find (&this->ignore_marker, guild_id);
-    if (l == this->ignore_marker.end ())
-        {
-            this->ignore_marker.push_back (guild_id);
-        }
-}
-
-void
-Manager::remove_ignore_marker (const dpp::snowflake &guild_id)
-{
-    std::lock_guard<std::mutex> lk (this->im_m);
-    auto l = vector_find (&this->ignore_marker, guild_id);
-    if (l != this->ignore_marker.end ())
-        {
-            this->ignore_marker.erase (l);
-        }
-}
-
-bool
-Manager::has_ignore_marker (const dpp::snowflake &guild_id)
-{
-    std::lock_guard<std::mutex> lk (this->im_m);
-    auto l = vector_find (&this->ignore_marker, guild_id);
-    if (l != this->ignore_marker.end ())
-        return true;
-    else
-        return false;
-}
-
-int
-Manager::load_guild_current_queue (const dpp::snowflake &guild_id,
-                                   const dpp::snowflake *user_id)
-{
-    auto player = this->create_player (guild_id);
-    if (player->saved_queue_loaded == true)
-        return 0;
-    player->saved_queue_loaded = true;
-
-    std::pair<PGresult *, ExecStatusType> res
-        = database::get_guild_current_queue (guild_id);
-
-    std::pair<std::deque<MCTrack>, int> queue
-        = database::get_playlist_from_PGresult (res.first);
-
-    database::finish_res (res.first);
-    res.first = nullptr;
-
-    if (queue.second != 0)
-        return queue.second;
-
-    for (auto &t : queue.first)
-        {
-            if (user_id)
-                t.user_id = *user_id;
-            player->add_track (t);
-        }
-
-    return queue.second;
-}
-
-int
-Manager::load_guild_player_config (const dpp::snowflake &guild_id)
-{
-    auto player = this->create_player (guild_id);
-    if (player->saved_config_loaded == true)
-        return 0;
-    player->saved_config_loaded = true;
-
-    std::pair<PGresult *, ExecStatusType> res
-        = database::get_guild_player_config (guild_id);
-
-    std::pair<database::player_config, int> conf
-        = database::parse_guild_player_config_PGresult (res.first);
-
-    database::finish_res (res.first);
-    res.first = nullptr;
-
-    if (conf.second != 0)
-        return conf.second;
-
-    player->loop_mode = conf.first.loop_mode;
-    player->max_history_size = (size_t)conf.first.autoplay_threshold;
-    player->auto_play = conf.first.autoplay_state;
-
-    return conf.second;
 }
 } // player
 } // musicat
